refactor: collapse @pixelrpg/data-gjs into @pixelrpg/ui-gjs/sprite
data-gjs was the GJS-side half of the old WebView-era data split: it
held the GObject-wrapped `GameProjectResource`, `SpriteSetResource`,
`SpriteSheet`, `Sprite`, `SpritePaintable` and `ImageResource` types
used exclusively by the editor UI to render sprite previews in
Gtk.Picture widgets. No other consumer exists now that the WebView
and its parallel web/WebKit targets are gone, so a standalone
package is not warranted.

- Move `packages/data-gjs/src/{objects,resource,utils,types}` into
  `packages/ui-gjs/src/sprite/`. Content is unchanged; only the
  package boundary disappears. Deleted the now-obsolete
  `packages/data-gjs/` directory in its entirety.
- Extract the build-time constants (`APPLICATION_ID`, `RESOURCES_PATH`,
  `PACKAGE_VERSION`, `PREFIX`, etc.) from `data-gjs/src/constants.ts`
  into a new `apps/maker-gjs/src/constants.ts` — they are app-level
  esbuild `define` values, not a shared package concern.
- Expose `@pixelrpg/ui-gjs/sprite` as an explicit package subpath in
  `packages/ui-gjs/package.json#exports`. Rewire every consumer:
  maker-gjs widgets (sidebar, project-view, application-window),
  ui-gjs widgets (sprite-sheet, tileset-selector, map-editor-panel,
  stories). Remove `@pixelrpg/data-gjs` from `apps/maker-gjs` and
  `packages/engine-gjs` workspace dependencies.
- Add `@girs/gsk-4.0: ^4.0.0-4.0.0-rc.3` directly to `engine-gjs`
  (data-gjs was silently pinning it at top level before; without it
  yarn resolved a stale `beta.40` that clashed with Gtk-4's required
  `rc.3`).
- Add `skipLibCheck: true` to `apps/storybook-gjs/tsconfig.json` —
  the dedup also exposed pre-existing duplicate `@girs/*` ambient
  declarations that were previously masked.
- Set `rootDir: ./src` in `packages/ui-gjs/tsconfig.json` to
  disambiguate the new `./sprite` exports entry for tsc.

Net: 46 files, +214 / −1792 LOC (−1578 LOC).
