+++ apps/maker-gjs/package.json
@@ -58,7 +58,6 @@
         "@gjsify/xmlhttprequest": "^0.1.11",
         "@gjsify/zlib": "^0.1.11",
         "@pixelrpg/data-excalibur": "workspace:^",
-        "@pixelrpg/data-gjs": "workspace:^",
         "@pixelrpg/engine-excalibur": "workspace:^",
         "@pixelrpg/engine-gjs": "workspace:^",
         "@pixelrpg/ui-gjs": "workspace:^",
