+++ packages/data-excalibur/src/index.ts
@@ -8,6 +8,13 @@ export * from "./format/SpriteSetFormat";
 export * from "./resource/GameProjectResource";
 export * from "./resource/MapResource";
 export * from "./resource/SpriteSetResource";
-export * from "./types/index";
+export * from "./types/GameStartupConfig";
+export * from "./types/Loadable";
+export * from "./types/MapCategory";
+export * from "./types/ResourceOptions";
+export * from "./types/Vector";
+export * from "./types/data/index";
+export * from "./types/reference/index";
+export * from "./types/resource/index";
 export * from "./utils/file";
 export * from "./utils/url";
