--- packages/data-gjs/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ESNext",
-        "module": "ESNext",
-        "moduleResolution": "bundler",
-        "esModuleInterop": true,
-        "strict": true,
-        "skipLibCheck": true,
-        "declaration": true,
-        "outDir": "dist",
-        "noEmit": true,
-        "allowImportingTsExtensions": true
-    },
-    "include": ["src/**/*", "env.d.ts"]
-}
