+++ packages/ui-gjs/src/sprite/index.ts
@@ -2,4 +2,3 @@ export * from './objects/index.ts'
 export * from './resource/index.ts'
 export * from './types/index.ts'
 export * from './utils/index.ts'
-export * from './constants.ts'
