+++ packages/ui-gjs/src/sprite/objects/Sprite.ts
