+++ packages/ui-gjs/src/sprite/objects/SpritePaintable.ts
