+++ packages/ui-gjs/src/sprite/objects/SpriteSheet.ts
