+++ packages/ui-gjs/src/sprite/objects/index.ts
