+++ packages/ui-gjs/src/sprite/resource/GameProjectResource.ts
