+++ packages/ui-gjs/src/sprite/resource/ImageResource.ts
