+++ packages/ui-gjs/src/sprite/resource/MapResource.ts
