+++ packages/ui-gjs/src/sprite/resource/SpriteSetResource.ts
