+++ packages/ui-gjs/src/sprite/resource/index.ts
