+++ packages/ui-gjs/src/sprite/types/GameProjectResourceOptions.ts
