+++ packages/ui-gjs/src/sprite/types/MapResourceOptions.ts
