+++ packages/ui-gjs/src/sprite/types/ResourceOptions.ts
