+++ packages/ui-gjs/src/sprite/types/SpriteSetResourceOptions.ts
