+++ packages/ui-gjs/src/sprite/types/index.ts
