+++ packages/ui-gjs/src/sprite/utils/file.ts
