+++ packages/ui-gjs/src/sprite/utils/index.ts
