+++ packages/ui-gjs/src/sprite/utils/path.ts
