+++ packages/ui-gjs/src/widgets/map-editor/tileset-selector.story.ts
@@ -9,7 +9,7 @@ import {
   StoryModule,
 } from '@pixelrpg/story-gjs'
 import { TilesetSelector } from './tileset-selector'
-import { SpriteSetResource, SpriteSheet } from '@pixelrpg/data-gjs'
+import { SpriteSetResource, SpriteSheet } from '@pixelrpg/ui-gjs/sprite'
 
 // Import story template
 import TilesetSelectorStoryTemplate from './tileset-selector.story.blp'
