+++ packages/ui-gjs/tsconfig.json
@@ -8,6 +8,7 @@
         "skipLibCheck": true,
         "declaration": true,
         "outDir": "dist",
+        "rootDir": "./src",
         "noEmit": true,
         "allowImportingTsExtensions": true,
         "resolveJsonModule": true
